pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "resumyst"
authors = [
    {name = "Filippo Guerranti"},
]
description = "Modern CV generation system combining Typst templating with Python CLI - create multiple professional document variants from unified YAML data"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup",
    "Topic :: Office/Business",
]
keywords = ["cv", "resume", "typst", "template", "cli", "yaml", "document-generation", "academic", "pdf", "professional"]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "watchdog>=3.0.0",
]
dynamic = ["version"]


[project.urls]
Homepage = "https://github.com/guerrantif/resumyst"
Repository = "https://github.com/guerrantif/resumyst"
Issues = "https://github.com/guerrantif/resumyst/issues"

[project.scripts]
resumyst = "resumyst:main"

[tool.flit.module]
name = "resumyst"